services/email/pyproject.toml
[tool.poetry]
name = "Email Service Anti-Freelancer"
version = "0.1.0"
description = ""
authors = ["_Counter021_"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {version = "0.68.1", extras = ["all"]}
celery = "^5.1.2"
redis = "^3.5.3"
PyJWT = "^2.3.0"
SQLAlchemy = {extras = ["asyncion"], version = "^1.4.26"}
asyncpg = "^0.24.0"
psycopg2-binary = "^2.9.1"
alembic = "^1.7.4"
aiohttp = "^3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
